Improved AI + scaling + drawing + time scaling.
We used a reverse A* algorithm and shared the board between all
enemies. They share information about which way is the most efficient
way to get to the player. They also cut corners by anticipating a turn
and moving towards the next next position.

I made the player follow the time curve as well as the bullets. It felt
better to have all elements in the game follow the same time curve.